add MORE_FLAGS flag
This is done in order to make it easier to add more flags in the future.

The MORE_FLAGS flag is always the least significant flag (located in
byte 1) and if present, there will be one more byte of flags.

If that byte also has the MORE_FLAGS flag, there will be another flag
byte and so on.

Signed-off-by: Erik Stromdahl <erik.stromdahl@gmail.com>